ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-multipart==0.0.6
requests==2.31.0
qdrant-client==1.7.0
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
langchain==0.0.345
sentence-transformers==2.2.2
torch==2.1.1
numpy==1.26.2
python-docx==1.1.0
pypdf==3.17.1
langchain-text-splitters==0.0.1
presidio-analyzer==2.2.351
presidio-anonymizer==2.2.351
